+++ main.py
@@ -1,79 +1,84 @@
+import time
 import pygame
+import menus
 import constants
 from lvl import Level01
-import time
 from bird import Bird
-import menus
-
-if not pygame.font:
-    print('Warning, fonts disabled')
-if not pygame.mixer:
-    print('Warning, sound disabled')
-
-pygame.init()
-pygame.font.init()
-
-# Define display and screen variables
-DISPLAY = (constants.SCREEN_WIDTH, constants.SCREEN_HEIGHT)
-screen = pygame.display.set_mode(DISPLAY)
-screen_rect = screen.get_rect()
-background = constants.BACKGROUND.convert()
-
-pygame.display.set_caption('Flappy Bird')
 
 
-def main():
+def main_menu_stage(main_menu: menus.MainMenu,
+                    screen: pygame.Surface) -> None:
+    while not main_menu.is_started:
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
+                raise SystemExit('QUIT')
+            elif event.type == pygame.MOUSEBUTTONUP and main_menu.is_hovered:
+                main_menu.is_started = True
+        main_menu.update()
+        main_menu.draw(screen)
+        pygame.display.flip()
+
+
+def get_ready_stage(get_ready_menu: menus.GetReadyMenu,
+                    screen: pygame.Surface, bird: Bird) -> None:
+    while not get_ready_menu.is_tapped:
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
+                raise SystemExit('QUIT')
+            elif event.type == pygame.KEYUP and event.key in (pygame.K_PAUSE, pygame.K_p):
+                paused = not paused
+            elif event.type == pygame.MOUSEBUTTONUP or (event.type == pygame.KEYUP and
+                                                        event.key in (pygame.K_UP,
+                                                        pygame.K_RETURN, pygame.K_SPACE)):
+                bird.jump
+                get_ready_menu.is_tapped = True
+        get_ready_menu.draw(screen)
+        pygame.display.flip()
+
+
+def game_over_stage(game_over_screen: menus.GameOverScreen,
+                    screen: pygame.Surface, current_level: Level01) -> None:
+    while not game_over_screen.restart:
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
+                raise SystemExit('QUIT')
+            elif event.type == pygame.MOUSEBUTTONUP and game_over_screen.is_hovered:
+                game_over_screen.restart = True
+        game_over_screen.update()
+        game_over_screen.draw(screen, current_level.score)
+        pygame.display.flip()
+
+
+def main() -> None:
+    pygame.init()
+    pygame.font.init()
+    DISPLAY = (constants.SCREEN_WIDTH, constants.SCREEN_HEIGHT)
+    screen = pygame.display.set_mode(DISPLAY)
+    background = constants.BACKGROUND.convert()
     bird = Bird(constants.BIRD_WIDTH, constants.BIRD_HEIGHT,
-                constants.BIRD_JUMP_SPEED, constants.BIRD_STARTING_POSITION,
+                constants.BIRD_JUMP_SPEED,
+                constants.BIRD_STARTING_POSITION,
                 constants.SFX_WING_PATH, constants.SFX_POINT_PATH,
                 constants.SFX_HIT_PATH, constants.SFX_DIE_PATH,
-                constants.BIRD_SPRITE_SHEET_PATH, constants.BIRD_FRAMES_COORDINATES)
+                constants.BIRD_SPRITE_SHEET_PATH,
+                constants.BIRD_FRAMES_COORDINATES)
 
     level_list = list()
     level_list.append(Level01(bird, background))
-
     current_level_no = 0
     current_level = level_list[current_level_no]
     current_level.create_level()
     bird.level = current_level
-    go = menus.GameOverScreen()
 
-    # Create and handle main menu
-    m = menus.MainMenu(bird, background, constants.MENU_ASSETS)
-    while not m.is_started:
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                raise SystemExit('QUIT')
-            elif event.type == pygame.MOUSEBUTTONUP and m.is_hovered:
-                m.is_started = True
-        m.update()
-        m.draw(screen)
-        pygame.display.flip()
+    clock = pygame.time.Clock()
+    main_menu = menus.MainMenu(bird, background, constants.MENU_ASSETS)
+    get_ready_menu = menus.GetReadyMenu(bird, background, constants.MENU_ASSETS)
+    game_over_screen = menus.GameOverScreen()
 
-    # Create pre game screen
-    pgs = menus.GetReadyMenu(bird, background, constants.MENU_ASSETS)
-
-    # Main game loop
+    # Game start
+    main_menu_stage(main_menu, screen)
     while True:
-
-        # Handle pre game screen inside the main loop
-        while not pgs.is_tapped:
-            for event in pygame.event.get():
-                if event.type == pygame.QUIT:
-                    raise SystemExit('QUIT')
-                elif event.type == pygame.KEYUP and event.key in (pygame.K_PAUSE, pygame.K_p):
-                    paused = not paused
-                elif event.type == pygame.MOUSEBUTTONUP or (event.type == pygame.KEYUP and
-                                                            event.key in (pygame.K_UP,
-                                                            pygame.K_RETURN, pygame.K_SPACE)):
-                    bird.jump
-                    start_time = time.time() - 4
-                    clock = pygame.time.Clock()
-                    pgs.is_tapped = True
-            pgs.draw(screen)
-            pygame.display.flip()
-
-        # Start of main game
+        get_ready_stage(get_ready_menu, screen, bird)
         for event in pygame.event.get():
             if event.type == pygame.QUIT:
                 raise SystemExit('QUIT')
@@ -82,8 +87,7 @@ def main():
             elif event.type == pygame.MOUSEBUTTONUP or (event.type == pygame.KEYUP and
                                                         event.key in (pygame.K_UP, pygame.K_RETURN, pygame.K_SPACE)):
                 bird.jump
-                start_time = time.time() - 4
-
+        start_time = time.time() - 4
         delta_time = start_time - time.time()
 
         current_level.update()
@@ -93,19 +97,13 @@ def main():
         bird.update(delta_time / 100)
 
         if bird.is_obstacle_hit:
-            go.update()
-            go.draw(screen, current_level.score)
-            while not go.restart:
-                for event in pygame.event.get():
-                    if event.type == pygame.QUIT:
-                        raise SystemExit('QUIT')
-                    elif event.type == pygame.MOUSEBUTTONUP and go.is_hovered:
-                        main()
-                pygame.display.flip()
+            game_over_stage(game_over_screen, screen, current_level)
+            break
 
         clock.tick(65)
         pygame.display.flip()
 
 
 if __name__ == '__main__':
-    main()
+    while True:
+        main()