+++ menus.py
@@ -32,13 +32,13 @@ class MainMenu(BaseMenu):
     def __init__(self, bird: Bird, background: pygame.Surface,
                  menu_assets_path: str) -> None:
         super().__init__(bird, background, menu_assets_path)
-        self.start_button = Image(constants.START_BUTTON_SPRITE_COORDS,
-                                  self.menu_assets)
-        self.start_button.rect = constants.START_BUTTON_POSITION
-        self.start_button_rect_object = constants.START_BUTTON_COORDINATES
-
-        self.logo = Image(constants.LOGO_SPRITE_COORDS, self.menu_assets)
-        self.logo.rect = constants.LOGO_POSITION
+        self.start_button = Image(self.menu_assets,
+                                  constants.START_BUTTON_SPRITE_COORDS,
+                                  constants.START_BUTTON_POSITION,
+                                  constants.START_BUTTON_COORDINATES)
+        self.logo = Image(self.menu_assets,
+                          constants.LOGO_SPRITE_COORDS,
+                          constants.LOGO_POSITION)
 
         self.sprite_list.add(self.start_button)
         self.sprite_list.add(self.logo)
@@ -48,7 +48,7 @@ class MainMenu(BaseMenu):
 
     def update(self):
         # Change hover status and cursor on hover over a start button
-        if self.start_button_rect_object.collidepoint(pygame.mouse.get_pos()):
+        if self.start_button.rect_object.collidepoint(pygame.mouse.get_pos()):
             pygame.mouse.set_cursor(*pygame.cursors.tri_left)
             self.is_hovered = True
         else:
@@ -60,13 +60,12 @@ class GetReadyMenu(BaseMenu):
     def __init__(self, bird: Bird, background: pygame.Surface,
                  menu_assets_path: str) -> None:
         super().__init__(bird, background, menu_assets_path)
-        self.get_ready_button = Image(constants.GET_READY_BUTTON_SPRITE_COORDS,
-                                      self.menu_assets)
-        self.get_ready_button.rect = constants.GET_READY_BUTTON_POSITION
-
-        self.tap_tip = Image(constants.TAP_TIP_SPRITE_COORDS,
-                             self.menu_assets)
-        self.tap_tip.rect = constants.TAP_TIP_POSITION
+        self.get_ready_button = Image(self.menu_assets,
+                                      constants.GET_READY_BUTTON_SPRITE_COORDS,
+                                      constants.GET_READY_BUTTON_POSITION)
+        self.tap_tip = Image(self.menu_assets,
+                             constants.TAP_TIP_SPRITE_COORDS,
+                             constants.TAP_TIP_POSITION)
 
         self.sprite_list.add(self.get_ready_button)
         self.sprite_list.add(self.tap_tip)
@@ -78,18 +77,16 @@ class GameOverScreen:
     def __init__(self) -> None:
         self.sprite_list = pygame.sprite.Group()
         self.menu_assets = constants.MENU_ASSETS
-        self.game_over_image = Image(constants.GAME_OVER_SPRITE_COORDS,
-                                     self.menu_assets)
-        self.game_over_image.rect = constants.GAME_OVER_POSITION
-
-        self.summary_banner = Image(constants.SUMMARY_SPRITE_COORD,
-                                    self.menu_assets)
-        self.summary_banner.rect = constants.SUMMARY_POSITION
-
-        self.ok_button = Image(constants.OK_BUTTON_SPRITE_COORDS,
-                               self.menu_assets)
-        self.ok_button.rect = constants.OK_BUTTON_POSITION
-        self.ok_button_rect_object = constants.OK_BUTTON_COORDINATES
+        self.game_over_image = Image(self.menu_assets,
+                                     constants.GAME_OVER_SPRITE_COORDS,
+                                     constants.GAME_OVER_POSITION)
+        self.summary_banner = Image(self.menu_assets,
+                                    constants.SUMMARY_SPRITE_COORD,
+                                    constants.SUMMARY_POSITION)
+        self.ok_button = Image(self.menu_assets,
+                               constants.OK_BUTTON_SPRITE_COORDS,
+                               constants.OK_BUTTON_POSITION,
+                               constants.OK_BUTTON_COORDINATES)
 
         self.sprite_list.add(self.game_over_image)
         self.sprite_list.add(self.summary_banner)
@@ -98,7 +95,11 @@ class GameOverScreen:
         self.is_hovered = False
         self.restart = False
 
-        self.nums = Numbers(large=False)
+        self.nums = Numbers(constants.NUMBERS_ASSETS,
+                            constants.BIG_NUMS_SPRITE_COORDS,
+                            constants.BIG_NUMS_POSITION,
+                            constants.SMALL_NUMS_SPRITE_COORDS,
+                            constants.SMALL_NUMS_POSITION)
 
     def draw(self, screen: pygame.Surface, score: int):
         self.sprite_list.draw(screen)
@@ -106,12 +107,11 @@ class GameOverScreen:
         score_sprites.draw(screen)
 
     def _generate_score_sprites(self, score):
-        score_sprites = pygame.sprite.Group()
-        score_sprites = self.nums.generate_score(score_sprites, score)
+        score_sprites = self.nums.generate_score(score, is_big=False)
         return score_sprites
 
     def update(self):
-        if self.ok_button_rect_object.collidepoint(pygame.mouse.get_pos()):
+        if self.ok_button.rect_object.collidepoint(pygame.mouse.get_pos()):
             pygame.mouse.set_cursor(*pygame.cursors.tri_left)
             self.is_hovered = True
         else: