+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.7.0] - 2019-08-11
+### Changed
+- Image class constructor requires rect and rect_object.
+- Refactor Numbers class.
+- Refactor main logic.
+
 ## [0.6.0] - 2019-08-10
 ### Changed
 - Refactor menus module.
@@ -11,5 +17,6 @@
 ### Added
 - Initial version.
 
-[0.6.0]: https://github.com/isylkin/flappy_bird_clone/compare/v0.5.0...v0.6.0
-[0.5.0]: https://github.com/isylkin/flappy_bird_clone/compare/1576c33...v0.5.0
+[0.7.0]: https://github.com/isylkin/flappy_bird_clone/compare/0.6.0...0.7.0
+[0.6.0]: https://github.com/isylkin/flappy_bird_clone/compare/0.5.0...0.6.0
+[0.5.0]: https://github.com/isylkin/flappy_bird_clone/compare/1576c33...0.5.0